refactor: split gomod collector as separate module
